+++ src/UIHelper.kt
@@ -1,8 +1,6 @@
 package org.aprsdroid.app
 
 import android.app.Activity
-import android.app.AlertDialog
-import android.app.ListActivity
 import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
@@ -12,7 +10,6 @@ import android.os.Environment
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
-import android.view.ContextMenu
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View